Final demo code.
Corrected waiting in between clicks on the driver, I noticed that there is a need first, to let the driver change the page (url) displayed before continuing the iteration, and second, to wait for every element that you want to scrap from it, i.e. wait for it to be visible to the driver.